pyproject.toml
[build-system]
requires = ["setuptools >= 40.6.0", "wheel", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "windIO"
description = "Frameworks defining the inputs and outputs for systems engineering MDAO of wind turbine and plants."
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "IEA Wind Task 37 and 55"},
#     { name = "", email = "" },
]
license = { file = "LICENSE" }
keywords = ["windIO"]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
    'jsonschema',
    'ruamel.yaml',
    'xarray',       # TODO: this is used to load netcdf files through the validation function XrResourceLoader
    "netcdf4",
]

dynamic = ["version"]

[project.optional-dependencies]
docs = [
    "sphinx",
    "pydata-sphinx-theme",
    'json-schema-for-humans',
    'gitpython',
    'sphinx_multiversion',

]
test = [
    'numpy',
    "pytest",
    "pytest-subtests",
    "pint",
    "scipy"
]

[tool.setuptools.packages.find]
exclude = [
    "docs*",
    "test*"
]

[tool.setuptools.package-data]
windIO = [
    "schemas/plant/*.yaml",     # Plant schema files
    "schemas/turbine/*.yaml",   # Turbine schema files
    "examples/**",              # windIO examples
]

[project.urls]
Homepage = "https://github.com/IEAWindSystems/windIO"
Documentation = "https://ieawindsystems.github.io/windio"

[tool.pytest.ini_options]
filterwarnings = [
    "error"
]

[tool.setuptools_scm]
version_file = "windIO/version.py"
